v1.0.0  15.01.22

+++ includes/class-bs-formular2-deactivator.php
@@ -30,7 +30,9 @@ class Bs_Formular2_Deactivator {
 	 * @since    1.0.0
 	 */
 	public static function deactivate() {
-        delete_option('bs-formular2_server_api');
+        if(BS_FORMULAR2_Requires_Activation) {
+            delete_option('bs-formular2_server_api');
+        }
         $infoTxt = 'deaktiviert am ' . date('d.m.Y H:i:s')."\r\n";
         file_put_contents(BS_FORMULAR2_PLUGIN_ADMIN_DIR . BS_FORMULAR2_BASENAME . 'license.txt',$infoTxt,  FILE_APPEND | LOCK_EX);
 	}
